Addressed second round of reviews
